requirements.txt
# Core dependencies
keyboard>=0.13.5
mss>=9.0.1
Pillow>=11.0.0
pytesseract>=0.3.10
requests>=2.31.0

# Optional: PySide6 for better UI (can use tkinter instead)
# PySide6>=6.6.1

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.3
pytest-mock>=3.12.0

# Type checking (optional)
mypy>=1.7.1
